% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_index.R
\name{abc_index}
\alias{abc_index}
\title{Identify accepted simulations}
\usage{
abc_index(parameters = param_reject)
}
\arguments{
\item{parameters}{accepted parameter values}
}
\value{
vector index of accepted simulations
}
\description{
This function returns an index of the identified accepted simulations
}
\author{
Joseph Lewis
}
\keyword{internal}
