% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_reject.R
\name{abc_reject}
\alias{abc_reject}
\title{Rejection sampling scheme for ABC}
\usage{
abc_reject(
  parameters = processed_params,
  lines = lines,
  summary_stat_target = 0,
  tol = 1
)
}
\arguments{
\item{parameters}{prior parameter values}

\item{lines}{SpatialLines}

\item{summary_stat_target}{a vector of the simulated summary statistic}

\item{tol}{tolerance, the required proportion of points accepted nearest the target values}
}
\value{
list of accepted parameter values
}
\description{
This function applies the rejection sampling scheme based on the efficacy of the summary statistic
}
\author{
Joseph Lewis
}
\keyword{internal}
