% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_rejection.R
\name{ABC_rejection}
\alias{ABC_rejection}
\title{Route Paths Rejection sampling scheme for ABC}
\usage{
ABC_rejection(
  input_data,
  model,
  priors,
  known_routes,
  validation = "euclidean",
  tol = NULL,
  cores = 1,
  output = "dataframe",
  drop_rows = FALSE
)
}
\arguments{
\item{input_data}{a list of input data to be used in the route modelling process}

\item{model}{an R function implementing the route model to be simulated.}

\item{priors}{a matrix of priors.}

\item{known_routes}{a Spatialknown_routes to be used when comparing against the simulated route paths}

\item{validation}{Method used to validate simulated routes against supplied line. Current implementations are: 'euclidean'}

\item{tol}{tolerance. If NULL returns all simulations}

\item{cores}{xx}

\item{output}{if 'matrix' (default) then a matrix of parameter values of the accepted simulations is returned. If 'routes' then all accepted simulated routes are returned with the parameter values attached as dataframe.}

\item{drop_rows}{if TRUE (default) then rejected simulations are dropped. If FALSE then all simulations retained.}
}
\value{
Matrix or Spatialknown_routesDataFrame dependent on output argument
}
\description{
This function launches a series of model simulations with model parameters drawn from the user-supplied prior values.
}
\details{
xxx
}
\author{
Joseph Lewis
}
