% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_parameters.R
\name{attach_parameters}
\alias{attach_parameters}
\title{Attach parameters dataframe to sf}
\usage{
attach_parameters(known_routes, parameters, ...)
}
\arguments{
\item{known_routes}{xxxx}

\item{parameters}{xxx}

\item{...}{see Details}
}
\value{
xxx
}
\description{
Attach parameters dataframe to sf
}
\author{
Joseph Lewis
}
