% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_routepaths.R
\name{bind_routepaths}
\alias{bind_routepaths}
\title{binds multiple routepath objects into one}
\usage{
bind_routepaths(x)
}
\arguments{
\item{x}{\code{list} of routepath objects}
}
\value{
\code{routepath}
}
\description{
binds multiple routepath objects into one
}
\author{
Joseph Lewis
}
\keyword{internal}
