% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Process prior parameter values}
\usage{
calculate_distance(routes, known_route, validation)
}
\arguments{
\item{routes}{simulated route paths}

\item{known_route}{a SpatialLines object of the known route to be used when comparing against the simulated route paths}

\item{validation}{Method to compare simulated route paths against known_routes}
}
\value{
Summary Statistic of distance from known route to simulated route
}
\description{
Process prior parameter values
}
\author{
Joseph Lewis
}
\keyword{internal}
