% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_routepath.R
\name{calculate_routepath}
\alias{calculate_routepath}
\title{Calculates routepath from origin and destination locations of sf lines using user-supplied conductance surface}
\usage{
calculate_routepath(x, locations)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{locations}{\code{sf points} origin and destination points}
}
\value{
\code{sf line} least-cost path from origin to destination
}
\description{
calculates the routepath from the origin and destination locations of lines using a user-supplied conductance surface
}
\author{
Joseph Lewis
}
\keyword{internal}
