% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_routepaths.R
\name{calculate_routepaths}
\alias{calculate_routepaths}
\title{Calculates route path from start and end locations of lines using user-supplied cost surface}
\usage{
calculate_routepaths(cost_surface, locations)
}
\arguments{
\item{cost_surface}{TransitionMatrix}

\item{locations}{Start and End SpatialPoints returned from extract_end_points()}
}
\value{
SpatialLines of route paths from start to end locations
}
\description{
This function calculates the route path from the start and end locations of lines using a user-supplied cost surface
}
\details{
xxx
}
\author{
Joseph Lewis
}
