% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_tolerance.R
\name{change_tolerance}
\alias{change_tolerance}
\title{Change tolerance of accepted simulated routes}
\usage{
change_tolerance(routes, tolerance, type = "absolute")
}
\arguments{
\item{routes}{simulated routes}

\item{tolerance}{Maximum tolerance value for routes to be accepted}

\item{type}{absolute or percentage}
}
\value{
routes
}
\description{
Change tolerance of accepted simulated routes
}
\author{
Joseph Lewis
}
