% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_route.R
\name{create_route}
\alias{create_route}
\title{Create a route using predefined parameter values}
\usage{
create_route(input_data, model, values, origin, destination)
}
\arguments{
\item{input_data}{a list of input data to be used in the route modelling process}

\item{model}{An R function implementing the route model to be used when creating a route}

\item{values}{a matrix of defined parameter values to be used in model function}

\item{origin}{xx}

\item{destination}{xx}
}
\value{
sf line
}
\description{
This function creates a route using predefined parameter values.
}
\details{
xxx
}
\author{
Joseph Lewis
}
