% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculate euclidean distance from simulated route path to known route}
\usage{
euclidean_distance(routes, known_route)
}
\arguments{
\item{known_route}{a SpatialLines object of the known route to be used when comparing against the simulated route paths}

\item{cost_surface}{Cost Surface denoting ease of traversing landscape}
}
\value{
Euclidean distance from simulated route path to known route
}
\description{
xxxx
}
\author{
Joseph Lewis
}
\keyword{internal}
