% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_end_points.R
\name{extract_end_points}
\alias{extract_end_points}
\title{Extract Start and End Points from known_route}
\usage{
extract_end_points(known_route)
}
\arguments{
\item{known_route}{sp Spatialknown_route}
}
\value{
sp points of Start and End of supplied known_route
}
\description{
This function extracts the start and end points from the supplied known_route.
}
\author{
Joseph Lewis
}
\keyword{internal}
