% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_priors.R
\name{extract_priors}
\alias{extract_priors}
\title{Extract parameter values from routepaths}
\usage{
extract_priors(routepaths)
}
\arguments{
\item{routepaths}{Extract parameter values from routepaths}

\item{lines}{routepaths}
}
\value{
parameter values
}
\description{
Extract parameter values from routepaths
}
\author{
Joseph Lewis
}
