% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_colnames.R
\name{fix_colnames}
\alias{fix_colnames}
\title{Fixes column names}
\usage{
fix_colnames(routepaths)
}
\arguments{
\item{routepaths}{simulated routepaths}
}
\value{
routepaths
}
\description{
Fixes column names
}
\author{
Joseph Lewis
}
