% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_distance.R
\name{frechet_distance}
\alias{frechet_distance}
\title{Calculate Frechet distance between simulated route path and known route}
\usage{
frechet_distance(routes, known_route)
}
\arguments{
\item{routes}{a SpatialLines object of the simulated route}

\item{known_route}{a SpatialLines object of the known route to be used when comparing against the simulated route paths}
}
\value{
Frechet distance between simulated route path and known route
}
\description{
Calculate Frechet distance between simulated route path and known route
}
\author{
Joseph Lewis
}
\keyword{internal}
