% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_weights_cs.R
\name{local_weights_cs}
\alias{local_weights_cs}
\title{Local Weight}
\usage{
local_weights_cs(cost_surface, p = 1, constrains = NULL, win, global_weight)
}
\arguments{
\item{cost_surface}{cost surface}

\item{p}{risk}

\item{constrains}{Boolean vector of adjacent cells not traversable}

\item{win}{Window}

\item{global_weight}{Importance of factor}
}
\value{
cost suface standardised to a maximum value of 1 and minimum value relative to 0 incorporating local range
}
\description{
Local Weight
}
\author{
Joseph Lewis
}
