% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_distance.R
\name{max_distance}
\alias{max_distance}
\title{Calculates route path from start and end locations of lines using user-supplied cost surface}
\usage{
max_distance(routepaths, lines, matrix)
}
\arguments{
\item{cost_surface}{TransitionMatrix}

\item{locations}{Start and End SpatialPoints returned from extract_end_points()}
}
\value{
SpatialLines of route paths from start to end locations
}
\description{
This function calculates the route path from the start and end locations of lines using a user-supplied cost surface
}
\author{
Joseph Lewis
}
\keyword{internal}
