% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_deviation_index.R
\name{path_deviation_index}
\alias{path_deviation_index}
\title{Calculate path deviation index from simulated route path to known route}
\usage{
path_deviation_index(routes, known_route)
}
\arguments{
\item{routes}{a SpatialLines object of the simulated route}

\item{known_route}{a SpatialLines object of the known route to be used when comparing against the simulated route paths}
}
\value{
Normalised Path Deviation Index of the simulated route path from the known route
}
\description{
Calculate path deviation index from simulated route path to known route
}
\author{
Joseph Lewis
}
\keyword{internal}
