% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{pdi_distance}
\alias{pdi_distance}
\title{Calculate path deviation index from simulated route path to known route}
\usage{
pdi_distance(route, known_route)
}
\arguments{
\item{route}{a SpatialLines object of the simulated route}

\item{known_route}{a SpatialLines object of the known route to be used when comparing against the simulated route paths}
}
\value{
Normalised Path Deviation Index of the simulated route path from the known route
}
\description{
Calculate path deviation index from simulated route path to known route
}
\author{
Joseph Lewis
}
\keyword{internal}
