% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GoF.R
\name{plot_GoF}
\alias{plot_GoF}
\title{Calculates route path from start and end locations of lines using user-supplied cost surface}
\usage{
plot_GoF(routepaths, thresholds, normalise = TRUE)
}
\arguments{
\item{routepaths}{routepaths of outputs from ABC rejection}

\item{normalise}{if TRUE (default) then percentage of routepaths within distance normalised to 1}

\item{threshold}{numeric values of distances between known route and simulated routes}
}
\value{
plot
}
\description{
This function calculates the route path from the start and end locations of lines using a user-supplied cost surface
}
\author{
Joseph Lewis
}
