% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_similarity.R
\name{plot_similarity}
\alias{plot_similarity}
\title{Calculates route path from start and end locations of lines using user-supplied cost surface}
\usage{
plot_similarity(matrix, thresholds, cluster = TRUE, ...)
}
\arguments{
\item{matrix}{matrix of outputs from ABC rejection}

\item{thresholdss}{numeric values of distances between known route and simulated routes}
}
\value{
plot
}
\description{
This function calculates the route path from the start and end locations of lines using a user-supplied cost surface
}
\author{
Joseph Lewis
}
