% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_cf.R
\name{post_cf}
\alias{post_cf}
\title{plot cost functions based on posterior parameter values}
\usage{
post_cf(routepaths, index, FUN, draws, type = "conductance", normalise = TRUE)
}
\arguments{
\item{routepaths}{\code{routepath}}

\item{index}{\code{numeric} index of posterior parameter values to be used in supplied function}

\item{FUN}{\code{function} function applied to prior parameter values}

\item{draws}{\code{numeric} Number of samples to draw from prior parameter values}

\item{type}{\code{type} 'conductance' (default) or 'cost'. if 'cost' then 1/conductance}

\item{normalise}{\code{logical} if TRUE (default) then cost functions scaled to between 0 and 1. if FALSE then cost functions not scaled}
}
\value{
\code{ggplot2} cost functions based on prior parameter values
}
\description{
plot cost functions based on posterior parameter values
}
\author{
Joseph Lewis
}
