% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_cf_check.R
\name{post_cf_check}
\alias{post_cf_check}
\title{plot cost functions based on posterior parameter values}
\usage{
post_cf_check(
  routepath,
  FUN,
  draws = 100,
  type = "conductance",
  normalise = TRUE,
  limits = 0.9,
  alpha = 1
)
}
\arguments{
\item{routepath}{\code{routepath}}

\item{FUN}{\code{function} function applied to prior parameter values. Supplied function arguments should be the same name as the prior matrix}

\item{draws}{\code{numeric} Number of samples to draw from posterior parameter values. Default is 100}

\item{type}{\code{type} 'conductance' (default) or 'cost'. if 'cost' then 1/conductance}

\item{normalise}{\code{logical} if TRUE (default) then cost functions scaled to between 0 and 1. if FALSE then cost functions not scaled}

\item{limits}{\code{numeric} limits of mathematical slope values supplied to FUN. This also controls the X axis limits of the plot}

\item{alpha}{\code{numeric} colour transparency of cost functions. Expected value from 0 to 1 (default is 1)}
}
\value{
\code{ggplot2} post cost functions
}
\description{
plot cost functions based on posterior parameter values
}
\author{
Joseph Lewis
}
