% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_route_check.R
\name{post_route_check}
\alias{post_route_check}
\title{simulate posterior route realisations based on a supplied routepath}
\usage{
post_route_check(
  input_data,
  model,
  routepath,
  known_routes,
  validation = "euclidean",
  summary_function = NULL,
  draws = 100,
  ncores = 1,
  spatial = TRUE
)
}
\arguments{
\item{input_data}{\code{list} Data used in the routepath ABC modelling procedure}

\item{model}{\code{function} Model to be used when simulating routes. This function requires that a \code{conductanceMatrix} is returned}

\item{routepath}{\code{routepath} routepath object containing the accepted parameter values}

\item{known_routes}{\code{sf line} Known routes to compare against the simulated routes}

\item{validation}{\code{character} Validation method used to assess fit of least-cost path against known route. Implemented methods include: 'euclidean' (Default), 'pdi', 'frechet', 'hausdorff'}

\item{summary_function}{\code{function} If function supplied (default NULL) then this function is used to summarise the distance measurements for all roads to a single value. if NULL then maximum distance is used (maximum deviation from known route)}

\item{draws}{\code{numeric} Number of samples to draw from posterior parameter values. Default is 100}

\item{ncores}{\code{numeric} Number of cores used during the routepath ABC procedure. Default value is 1}

\item{spatial}{\code{logical} if TRUE (default) then sf object returned. if FALSE then sf geometry is dropped and a data.frame is returned}
}
\value{
\code{routepath}
}
\description{
simulate posterior route realisations based on a supplied routepath
}
\details{
xxx
}
\author{
Joseph Lewis
}
