% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_cf_check.R
\name{pp_cf_check}
\alias{pp_cf_check}
\title{prior and predictive check}
\usage{
pp_cf_check(priors, distribution = "laplace", no_samples = 100, alpha = 0.1)
}
\arguments{
\item{priors}{a matrix or data frame of priors.}

\item{distribution}{cost factor distribution. Currently only implemented the 'laplace' (default)}

\item{no_samples}{number of samples drawn from distribution}
}
\value{
xx
}
\description{
This function calculates the route path from the start and end locations of lines using a user-supplied cost surface
}
\author{
Joseph Lewis
}
