% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_cf.R
\name{prior_cf}
\alias{prior_cf}
\title{plot cost functions based on prior parameter values}
\usage{
prior_cf(priors, index, FUN, draws, type = "conductance", normalise = TRUE)
}
\arguments{
\item{priors}{\code{Matrix} prior parameter values}

\item{index}{\code{numeric} index of prior parameter values to be used in supplied function}

\item{FUN}{\code{function} function applied to prior parameter values}

\item{draws}{\code{numeric} Number of samples to draw from prior parameter values}

\item{type}{\code{type} 'conductance' (default) or 'cost'. if 'cost' then 1/conductance}

\item{normalise}{\code{logical} if TRUE (default) then cost functions scaled to between 0 and 1. if FALSE then cost functions not scaled}
}
\value{
\code{ggplot2} cost functions based on prior parameter values
}
\description{
plot cost functions based on prior parameter values
}
\examples{

priors <- cbind(a = 1, b = rnorm(n = 100, mean = 3.5, sd = 1),
c = rnorm(n = 100, mean = 0.05, sd = 0.2))
FUN <- function(x,a,b,c) {(a * exp(-b * abs(x + c))) / 3.6}

prior_cf(priors = priors, FUN = FUN, index = 1:3, draws = 10,
type = "conductance", normalise = TRUE)
}
\author{
Joseph Lewis
}
