% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_cf_check.R
\name{prior_cf_check}
\alias{prior_cf_check}
\title{plot cost function realisations based on supplied prior parameter values and a function}
\usage{
prior_cf_check(
  priors,
  FUN,
  draws = 100,
  type = "conductance",
  normalise = TRUE,
  limits = 0.9,
  alpha = 1
)
}
\arguments{
\item{priors}{\code{Matrix} prior parameter values}

\item{FUN}{\code{function} function applied to prior parameter values. Supplied function arguments should be the same name as the prior matrix}

\item{draws}{\code{numeric} Number of samples to draw from prior parameter values. Default is 100}

\item{type}{\code{type} 'conductance' (default) or 'cost'. if 'cost' then 1/conductance}

\item{normalise}{\code{logical} if TRUE (default) then cost functions scaled to between 0 and 1. if FALSE then cost functions not scaled}

\item{limits}{\code{numeric} limits of mathematical slope values supplied to FUN. This also controls the X axis limits of the plot}

\item{alpha}{\code{numeric} colour transparency of cost functions. Expected value from 0 to 1 (default is 1)}
}
\value{
\code{ggplot2} prior cost functions
}
\description{
plot cost function realisations based on supplied prior parameter values and a function
}
\details{
xxx
}
\examples{

priors <- cbind(a = 1, b = rnorm(n = 100, mean = 3.5, sd = 1),
c = rnorm(n = 100, mean = 0.05, sd = 0.2))
FUN <- function(x,a,b,c) {(a * exp(-b * abs(x + c))) / 3.6}

prior_cf_check(priors = priors, FUN = FUN, draws = 10,
type = "conductance", normalise = TRUE)
}
\author{
Joseph Lewis
}
