% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_route_check.R
\name{prior_route_check}
\alias{prior_route_check}
\title{simulate prior route realisations based on supplied prior parameter values and a model}
\usage{
prior_route_check(input_data, model, priors, known_routes, ncores = 1)
}
\arguments{
\item{input_data}{\code{list} Data used in the routepath ABC modelling procedure}

\item{model}{\code{function} Model to be used when simulating routes. This function requires that a \code{conductanceMatrix} is returned}

\item{priors}{\code{Matrix} Prior parameter values}

\item{known_routes}{\code{sf line} Known routes to compare against the simulated routes}

\item{ncores}{\code{numeric} Number of cores used during the routepath ABC procedure. Default value is 1}
}
\value{
\code{sf line} simulated routes
}
\description{
simulate prior route realisations based on supplied prior parameter values and a model
}
\details{
xxx
}
\author{
Joseph Lewis
}
