% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_abc.R
\name{process_abc}
\alias{process_abc}
\title{Process parameter values from Rejection sampling scheme for ABC}
\usage{
process_abc(parameters = param_reject, lines = lines)
}
\arguments{
\item{parameters}{accepted parameter values}

\item{lines}{SpatialLines}
}
\value{
Matrix of accepted prior parameter values
}
\description{
Process parameter values from Rejection sampling scheme for ABC
}
\author{
Joseph Lewis
}
\keyword{internal}
