% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_parameters.R
\name{process_parameters}
\alias{process_parameters}
\title{Process prior parameter values}
\usage{
process_parameters(
  routepaths,
  priors,
  line_id,
  row_no = row_no,
  summary_stat,
  spatial = spatial
)
}
\arguments{
\item{routepaths}{simulated route paths}

\item{priors}{a matrix or dataframe of priors}

\item{line_id}{ID attached to output. Default value is 1. Useful if iteratively modelling routes and want to incrementally assign line IDs.}

\item{row_no}{Row number}

\item{summary_stat}{Summary Statistic of distance from known route to simulated route}

\item{spatial}{if TRUE then sf Lines returned. If FALSE then dataframe returned}
}
\value{
routepaths. If Spatial TRUE then sf. If Spatial FALSE then dataframe
}
\description{
Process prior parameter values
}
\author{
Joseph Lewis
}
\keyword{internal}
