% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_route.R
\name{process_route}
\alias{process_route}
\title{Process route}
\usage{
process_route(
  route,
  priors,
  line_id,
  row_no = row_no,
  summary_stat,
  spatial = spatial
)
}
\arguments{
\item{route}{\code{sf line} least-cost path}

\item{priors}{\code{Matrix} prior parameter values to be included within routepath object}

\item{line_id}{\code{numeric} value to be included within routepath object}

\item{row_no}{\code{numeric} value to be included within the routepath object. This value corresponds to the row of the prior parameter Matrix. Default is a value of 1}

\item{summary_stat}{\code{numeric} distance betwen simulated route and known route to be included within routepath object}

\item{spatial}{\code{logical} if TRUE then sf object returned. if FALSE (default) then data.frame object returned}
}
\value{
\code{sf object} or \code{data.frame}
}
\description{
Process route
}
\author{
Joseph Lewis
}
\keyword{internal}
