% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_cs_global.R
\name{rescale_cs_global}
\alias{rescale_cs_global}
\title{rescale conductanceMatrix based on global range of values}
\usage{
rescale_cs_global(x, FUN = NULL)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{FUN}{\code{function} function applied to standardised conductanceMatrix values}
}
\value{
\code{conductanceMatrix} standardised conductanceMatrix with a maximum value of 1 and minimum value relative to 0
}
\description{
rescale conductanceMatrix based on global range of values
}
\author{
Joseph Lewis
}
