% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_cs_local.R
\name{rescale_cs_local}
\alias{rescale_cs_local}
\title{Rescales cost surface based on local range of transition values}
\usage{
rescale_cs_local(cost_surface, p = 1, constrains = NULL, win)
}
\arguments{
\item{cost_surface}{cost surface}

\item{p}{risk}

\item{constrains}{Boolean vector of adjacent cells not traversable}

\item{win}{Window}
}
\value{
cost suface standardised to a maximum value of 1 and minimum value relative to 0 incorporating local range
}
\description{
Rescales cost surface based on local range of transition values
}
\author{
Joseph Lewis
}
