% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_hists.R
\name{route_hists}
\alias{route_hists}
\title{plot scatter plot of simulated routepath parameters}
\usage{
route_hists(routes, col_index = NULL, xlab_names = NULL, ...)
}
\arguments{
\item{routes}{simulated routes}

\item{col_index}{column index of parameters to plot. if NULL then plot all parameters}

\item{xlab_names}{xxxx}

\item{...}{Additional arguments. See Details}
}
\description{
plot scatter plot of simulated routepath parameters
}
\author{
Joseph Lewis
}
