% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_pairs.R
\name{route_pairs}
\alias{route_pairs}
\title{plot pair plot of simulated routepath parameters}
\usage{
route_pairs(routes, col_index = NULL, xlab_names = NULL, alpha = 1)
}
\arguments{
\item{routes}{simulated routes}

\item{col_index}{column index of parameters to plot. if NULL then plot all parameters}

\item{xlab_names}{xxxx}

\item{alpha}{xxxx}
}
\description{
plot pair plot of simulated routepath parameters
}
\author{
Joseph Lewis
}
