% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_params.R
\name{route_params}
\alias{route_params}
\title{plot summary of simulated routepath parameters}
\usage{
route_params(routes, col_index = NULL, filename = NULL, xlab_names = NULL, ...)
}
\arguments{
\item{routes}{simulated routes}

\item{col_index}{column index of parameters to plot. if NULL then plot all parameters}

\item{filename}{xxx}

\item{xlab_names}{xxxx}

\item{...}{Additional arguments. See Details}
}
\description{
plot summary of simulated routepath parameters
}
\author{
Joseph Lewis
}
