% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_stats.R
\name{route_stats}
\alias{route_stats}
\title{plot summary of simulated routepath stats}
\usage{
route_stats(routes, filename = NULL, ...)
}
\arguments{
\item{routes}{simulated routes}

\item{filename}{xxxxx}

\item{...}{Additional arguments. See Details}
}
\description{
plot summary of simulated routepath stats
}
\author{
Joseph Lewis
}
