% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_parameters.R
\name{sample_parameters}
\alias{sample_parameters}
\title{Extract parameter values from routepaths}
\usage{
sample_parameters(parameters, no_samples)
}
\arguments{
\item{parameters}{Extract parameter values from routepaths}

\item{no_samples}{routepaths}
}
\value{
parameter values
}
\description{
Extract parameter values from routepaths
}
\author{
Joseph Lewis
}
