% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_route.R
\name{simulate_routes}
\alias{simulate_routes}
\title{simulate route realisations based on supplied parameter values and a model}
\usage{
simulate_routes(input_data, model, values, known_routes, ncores = 1)
}
\arguments{
\item{input_data}{\code{list} Data used in the routepath ABC modelling procedure}

\item{model}{\code{function} Model to be used when simulating routes. This function requires that a \code{conductanceMatrix} is returned}

\item{values}{\code{Matrix} Parameter values to be used in the model. Only the first row of the Matrix is used}

\item{known_routes}{\code{sf line} Known routes used for origin and destinations of simulated routes#}

\item{ncores}{\code{numeric} Number of cores used during the routepath ABC procedure. Default value is 1}
}
\value{
\code{sf line} simulated routes
}
\description{
simulate route realisations based on supplied parameter values and a model
}
\details{
xxx
}
\author{
Joseph Lewis
}
