% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_priors.R
\name{standardise_priors}
\alias{standardise_priors}
\title{Standardise priors}
\usage{
standardise_priors(priors, col_index)
}
\arguments{
\item{priors}{matrix of prior parameter values}

\item{col_index}{Column index of prior parameter values to standardise}
}
\value{
priors matrix
}
\description{
Priors in selected columns standardised to sum to one
}
\author{
Joseph Lewis
}
