% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_parameters.R
\name{summarise_parameters}
\alias{summarise_parameters}
\title{Summarise parameters}
\usage{
summarise_parameters(routes, col_index, fun)
}
\arguments{
\item{routes}{simulated routes}

\item{col_index}{...}

\item{fun}{....}
}
\value{
df
}
\description{
Summarise parameters
}
\author{
Joseph Lewis
}
