% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tolerance.R
\name{update_tolerance}
\alias{update_tolerance}
\title{update tolerance}
\usage{
update_tolerance(routepath, tol, type = "absolute")
}
\arguments{
\item{routepath}{\code{routepath}}

\item{tol}{\code{numeric} maximum distance between simulated routes and known route to be deemed equal}

\item{type}{\code{character} 'abolute' or 'percentage'. if 'absolute' (default) then all simulated routes above this value are rejected. if 'percentage' then all simulated routes with a stat above the supplied percentage are rejected}
}
\value{
\code{routepath}
}
\description{
Updates the maximum distance between simulated routes and known route to be deemed equal. All simulated routes above this value are rejected
}
\author{
Joseph Lewis
}
\keyword{internal}
