% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_rejection.R
\name{ABC_rejection}
\alias{ABC_rejection}
\title{routepath rejection sampling scheme using Approximate Bayesian Computation}
\usage{
ABC_rejection(
  input_data,
  model,
  priors,
  known_routes,
  validation = "euclidean",
  summary_function = NULL,
  tol = NULL,
  ncores = 1,
  spatial = TRUE
)
}
\arguments{
\item{input_data}{\code{list} Data used in the routepath ABC modelling procedure}

\item{model}{\code{function} Model to be used when simulating routes. This function requires that a \code{conductanceMatrix} is returned}

\item{priors}{\code{Matrix} Prior parameter values}

\item{known_routes}{\code{sf line} Known routes to compare against the simulated routes}

\item{validation}{\code{character} Validation method used to assess fit of least-cost path against known route. Implemented methods include: 'euclidean' (Default), 'pdi', 'frechet', 'hausdorff'}

\item{summary_function}{\code{function} If function supplied then this function is used to summarise the distance measurements for all modelled roads to a single value. Default is NULL}

\item{tol}{\code{numeric} Maximum distance between simulated routes and the known route for the two lines to still be deemed equal. All simulated routes and their parameters that result in a simulated route with a maximum distance from the known route above this value are rejected}

\item{ncores}{\code{numeric} Number of cores used during the routepath ABC procedure. Default value is 1}

\item{spatial}{\code{logical} if TRUE (default) then sf object returned. if FALSE then sf geometry is dropped and a data.frame is returned}
}
\value{
\code{routepath}
}
\description{
Launches a series of model simulations using user-supplied prior parameter values. See details for more information
}
\details{
xxx
}
\author{
Joseph Lewis
}
