% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_rejection.R
\name{ABC_rejection}
\alias{ABC_rejection}
\title{Routepath Rejection sampling scheme using Approximate Bayesian Computation}
\usage{
ABC_rejection(
  input_data,
  model,
  priors,
  known_route,
  validation = "euclidean",
  tol = NULL,
  cores = 1,
  spatial = FALSE,
  drop_rows = FALSE,
  line_id = 1
)
}
\arguments{
\item{input_data}{a list of input data to be used in the routepath modelling process}

\item{model}{An R function implementing the route model to be simulated. This function is expected to return a TransitionLayer object}

\item{priors}{a matrix of priors}

\item{known_route}{a SpatialLines object of the known route to compare against the simulated routepaths}

\item{validation}{Distance method used to validate simulated routes against the supplied known route. Current implementations are: 'euclidean', 'pdi', 'frechet' and 'hausdorff'}

\item{tol}{Tolerance threshold. Maximum deviation from known route for the simulated routepaths to be accepted. If NULL all simulated routepaths are returned.}

\item{cores}{Number of cores}

\item{spatial}{if TRUE then sf Lines returned. If FALSE then dataframe returned}

\item{drop_rows}{if FALSE (default) all simulated routepaths returned. If TRUE rejected simulations are dropped}

\item{line_id}{ID attached to routepaths output. Default value is 1. Useful if iteratively modelling known routes and want to incrementally assign line IDs}
}
\value{
Dataframe or sf Lines
}
\description{
This function launches a series of model simulations using user-supplied prior parameter values
}
\details{
xxx
}
\author{
Joseph Lewis
}
