% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_rejection.R
\name{ABC_rejection}
\alias{ABC_rejection}
\title{Route Paths Rejection sampling scheme for Approximate Bayesian Computation}
\usage{
ABC_rejection(
  input_data,
  model,
  priors,
  known_route,
  validation = "euclidean",
  tol = NULL,
  cores = 1,
  spatial = FALSE,
  drop_rows = FALSE,
  line_id = 1
)
}
\arguments{
\item{input_data}{a list of input data to be used in the route path modelling process}

\item{model}{An R function implementing the route model to be simulated.}

\item{priors}{a matrix or dataframe of priors}

\item{known_route}{a SpatialLines object of the known route to be used when comparing against the simulated route paths}

\item{validation}{Method used to validate simulated routes against supplied line. Current implementations are: 'euclidean', 'pdi' and 'frechet'}

\item{tol}{tolerance. Maximum deviation from known route for the simulation to be accepted. If NULL all simulated route paths are returned.}

\item{cores}{Number of cores}

\item{spatial}{if TRUE then sf Lines returned. If FALSE then dataframe returned}

\item{drop_rows}{if TRUE (default) then rejected simulations are dropped. If FALSE then all simulations returned}

\item{line_id}{ID attached to output. Default value is 1. Useful if iteratively modelling routes and want to incrementally assign line IDs}
}
\value{
Dataframe or sf Lines
}
\description{
This function launches a series of model simulations with model parameters drawn from the user-supplied prior values.
}
\details{
xxx
}
\author{
Joseph Lewis
}
