% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABC_rejection.R
\name{ABC_rejection}
\alias{ABC_rejection}
\title{Route Paths Rejection sampling scheme for ABC}
\usage{
ABC_rejection(
  input_data,
  model,
  priors,
  lines,
  validation = "max_distance",
  summary_stat_target = 0,
  tol = 1,
  cores = 1,
  output = "matrix"
)
}
\arguments{
\item{input_data}{a list of input data to be used in the route modelling process}

\item{model}{an R function implementing the route model to be simulated.}

\item{priors}{a matrix of priors.}

\item{lines}{a SpatialLines to be used when comparing against the simulated route paths}

\item{validation}{Method used to validate simulated routes against supplied line. Current implementations are: 'max_distance'}

\item{summary_stat_target}{a vector containing the targeted (observed) summary statistics.}

\item{tol}{tolerance, a strictly positive number (between 0 and 1) indicating the proportion of simulations retained nearest the targeted summary statistics.}

\item{cores}{xx}

\item{output}{if "matrix" (default) then a matrix of parameter values of the accepted simulations is returned. If "routes" then all accepted simulated routes are returned with the parameter values attached as dataframe.}

\item{sim_routes}{if TRUE simulated route paths are also returned}
}
\value{
Matrix or SpatialLinesDataFrame dependent on output argument
}
\description{
This function launches a series of model simulations with model parameters drawn from the user-supplied prior values.
}
\details{
xxx
}
\author{
Joseph Lewis
}
