% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Process prior parameter values}
\usage{
calculate_distance(routes, known_routes, validation)
}
\arguments{
\item{known_routes}{Spatialknown_routes}

\item{validation}{Method to compare simulated route paths against known_routes}

\item{routepaths}{simulated route paths}

\item{matrix}{matrix}
}
\value{
summary statistic
}
\description{
Process prior parameter values
}
\author{
Joseph Lewis
}
\keyword{internal}
