% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{calculates distance between simulated route and known route}
\usage{
calculate_distance(route, known_route, validation)
}
\arguments{
\item{route}{\code{sf line} least-cost path}

\item{known_route}{\code{sf line} known route}

\item{validation}{\code{character} validation method used to assess fit of least-cost path against known route. Implemented methods include: 'euclidean' (Default), 'pdi', 'frechet', 'hausdorff'}
}
\value{
distance betwen simulated route and known route
}
\description{
Calculates distance between simulated route and known route using the validation method chosen
}
\author{
Joseph Lewis
}
\keyword{internal}
