% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_routepath.R
\name{calculate_routepath}
\alias{calculate_routepath}
\title{Calculates route path from start and end locations of lines using user-supplied cost surface}
\usage{
calculate_routepath(x, locations)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{locations}{\code{sf points} origin and destination points}
}
\value{
\code{sf line} least-cost path from origin to destination
}
\description{
calculates the route path from the start and end locations of lines using a user-supplied cost surface
}
\author{
Joseph Lewis
}
\keyword{internal}
