% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_tolerance.R
\name{change_tolerance}
\alias{change_tolerance}
\title{Change tolerance of accepted simulated routepaths}
\usage{
change_tolerance(routepaths, tolerance)
}
\arguments{
\item{routepaths}{simulated routepaths}

\item{tolerance}{Maximum tolerance value for routepaths to be accepted}
}
\value{
routepaths
}
\description{
Change tolerance of accepted simulated routepaths
}
\author{
Joseph Lewis
}
