% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_distance.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculates route path from start and end locations of known_routes using user-supplied cost surface}
\usage{
euclidean_distance(routes, known_routes)
}
\arguments{
\item{cost_surface}{TransitionMatrix}

\item{locations}{Start and End SpatialPoints returned from extract_end_points()}
}
\value{
Spatialknown_routes of route paths from start to end locations
}
\description{
This function calculates the route path from the start and end locations of known_routes using a user-supplied cost surface
}
\author{
Joseph Lewis
}
\keyword{internal}
