% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Calculate euclidean distance from simulated route path to known route}
\usage{
euclidean_distance(route, known_route)
}
\arguments{
\item{route}{a sf object of the simulated route}

\item{known_route}{a sf object of the known route to be used when comparing against the simulated route paths}
}
\value{
Euclidean distance from simulated route path to known route
}
\description{
Calculate euclidean distance from simulated route path to known route
}
\author{
Joseph Lewis
}
\keyword{internal}
