% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_end_points.R
\name{extract_end_points}
\alias{extract_end_points}
\title{Extracts origin and destination locations from known_route}
\usage{
extract_end_points(known_route)
}
\arguments{
\item{known_route}{\code{sf line} known route}
}
\value{
sf POINT origin and destination of supplied known_route
}
\description{
Extracts the origin and destination point from the supplied known route.
}
\author{
Joseph Lewis
}
\keyword{internal}
