% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_end_points.R
\name{extract_end_points}
\alias{extract_end_points}
\title{Extract Start and End Points from known_routes}
\usage{
extract_end_points(known_routes)
}
\arguments{
\item{known_routes}{sp Spatialknown_routes}
}
\value{
sp points of Start and End of supplied known_routes
}
\description{
This function extracts the start and end points from the supplied known_routes.
}
\author{
Joseph Lewis
}
\keyword{internal}
