% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_end_points.R
\name{extract_end_points}
\alias{extract_end_points}
\title{Extract Start and End Points from Lines}
\usage{
extract_end_points(lines)
}
\arguments{
\item{lines}{sp SpatialLines}
}
\value{
sp points of Start and End of supplied lines
}
\description{
This function extracts the start and end points from the supplied lines.
}
\author{
Joseph Lewis
}
\keyword{internal}
