% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_wlc.R
\name{global_wlc}
\alias{global_wlc}
\title{Global Weighted Linear Combination}
\usage{
global_wlc(cost_surface, p = 1, constrains = NULL, global_weight)
}
\arguments{
\item{cost_surface}{cost surface}

\item{p}{risk}

\item{constrains}{Boolean vector of adjacent cells not traversable}

\item{global_weight}{Importance of factor}
}
\value{
Global Weighted Linear Combination
}
\description{
Global Weighted Linear Combination
}
\author{
Joseph Lewis
}
