% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_wlc.R
\name{global_wlc}
\alias{global_wlc}
\title{Global Weighted Linear Combination}
\usage{
global_wlc(
  cost_surface,
  constrains = NULL,
  global_weight,
  neighbours,
  fun = NULL
)
}
\arguments{
\item{cost_surface}{cost surface}

\item{constrains}{Boolean vector of adjacent cells not traversable}

\item{global_weight}{Importance of factor}

\item{neighbours}{xxx}

\item{fun}{An R function to apply to the standardised transition values}
}
\value{
Global Weighted Linear Combination
}
\description{
Global Weighted Linear Combination
}
\author{
Joseph Lewis
}
