% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_wlc.R
\name{global_wlc}
\alias{global_wlc}
\title{Global Weighted Linear Combination}
\usage{
global_wlc(x, global_weight, FUN = NULL)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{global_weight}{\code{numeric} importance of factor. Value must be between 0 and 1}

\item{FUN}{\code{function} function applied to standardised conductanceMatrix values}
}
\value{
\code{conductanceMatrix}
}
\description{
Global Weighted Linear Combination
}
\author{
Joseph Lewis
}
