% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_weights_cs.R
\name{local_weights_cs}
\alias{local_weights_cs}
\title{rescale conductanceMatrix based on global range of values}
\usage{
local_weights_cs(x, global_weight)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{global_weight}{\code{numeric} importance of factor. Supplied value must be between 0 and 1}
}
\description{
rescale conductanceMatrix based on global range of values
}
