% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_wlc.R
\name{local_wlc}
\alias{local_wlc}
\title{Local Weighted Linear Combination}
\usage{
local_wlc(
  cost_surface,
  constrains = NULL,
  global_weight,
  neighbours,
  fun = NULL
)
}
\arguments{
\item{cost_surface}{cost surface}

\item{constrains}{Boolean vector of adjacent cells not traversable}

\item{global_weight}{Importance of factor}

\item{neighbours}{xxx}

\item{fun}{xxx}
}
\value{
Local Weighted Linear Combination
}
\description{
Local Weighted Linear Combination
}
\author{
Joseph Lewis
}
