% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_wlc.R
\name{local_wlc}
\alias{local_wlc}
\title{Local Weighted Linear Combination}
\usage{
local_wlc(cost_surface, p = 1, constrains = NULL, win, global_weight)
}
\arguments{
\item{cost_surface}{cost surface}

\item{p}{risk}

\item{constrains}{Boolean vector of adjacent cells not traversable}

\item{win}{Window}

\item{global_weight}{Importance of factor}
}
\value{
Local Weighted Linear Combination
}
\description{
Local Weighted Linear Combination
}
\author{
Joseph Lewis
}
