% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_parameters.R
\name{process_parameters}
\alias{process_parameters}
\title{Process prior parameter values}
\usage{
process_parameters(routepaths, lines, priors, validation)
}
\arguments{
\item{routepaths}{simulated route paths}

\item{lines}{SpatialLines}

\item{priors}{prior parameter values}

\item{validation}{Method to compare simulated route paths against lines}
}
\value{
list of prior parameter values
}
\description{
Process prior parameter values
}
\author{
Joseph Lewis
}
\keyword{internal}
