% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_parameters.R
\name{process_parameters}
\alias{process_parameters}
\title{Process prior parameter values}
\usage{
process_parameters(
  routepaths,
  known_routes,
  priors,
  summary_stat,
  output = output,
  row_no = row_no
)
}
\arguments{
\item{routepaths}{simulated route paths}

\item{known_routes}{Spatialknown_routes}

\item{priors}{prior parameter values}

\item{output}{Method to compare simulated route paths against known_routes}

\item{validation}{Method to compare simulated route paths against known_routes}
}
\value{
list of prior parameter values
}
\description{
Process prior parameter values
}
\author{
Joseph Lewis
}
\keyword{internal}
