% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_cs_local.R
\name{rescale_cs_local}
\alias{rescale_cs_local}
\title{rescale conductanceMatrix based on global range of values}
\usage{
rescale_cs_local(x, FUN = NULL)
}
\arguments{
\item{x}{\code{conductanceMatrix}}

\item{FUN}{\code{function} function applied to standardised conductanceMatrix values}
}
\value{
\code{conductanceMatrix}
}
\description{
rescale conductanceMatrix based on global range of values
}
\author{
Joseph Lewis
}
